Update build-system requirement for setuptools-scm to >=7.0.0
pyproject.toml:
Update build-system requirement for setuptools-scm to >=7.0.0, which
obsoletes the use of setuptools-scm-git-archive.
Remove pip and wheel from build-system as they are not required in a
PEP517 build setup using pypa/build and pypa/installer.

setup.{cfg,py}:
Update build-system requirement for setuptools-scm to >=7.0.0, which
obsoletes the use of setuptools-scm-git-archive.