Don't strip newlines when ANSI text is sent to print()
AnsiDecoder.decode() takes a single str and returns a list of Text split
by newline. If this is used directly, newlines in the source will not be
restored when sent to print().

Using Lines instead of list lets us indicate to Rich that these Text
objects should be interspersed with newlines.

Ref https://github.com/Textualize/rich/blob/master/rich/ansi.py#L126